matrices

+++ usefulprograms/logarithm.java
@@ -4,14 +4,14 @@ public class logarithm {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		System.out.println("Enter the base: ");
-		int base = scan.nextInt();
+		double base = scan.nextDouble();
 		System.out.println("Enter the number: ");
 		double number = scan.nextDouble();
 		System.out.println("Log" + base + "    " + number + "   =   " + log(base, number));
 	}
-	public static double log(int base, double number) {
+	public static double log(double base, double number) {
 		double power = 1;
-		while (base != (int)number) {
+		while (base != number) {
 			number /= base;
 			power++;
 		}
